+++ src/cljs-lib/src/calva/fmt/inferer.cljs
@@ -1,5 +1,5 @@
 (ns calva.fmt.inferer
-  (:require ["parinfer" :as parinfer]
+  (:require ["@chrisoakman/parinfer" :as parinfer]
             [calva.js-utils :refer [cljify jsify]]
             [calva.fmt.editor :as editor]))
 
@@ -10,16 +10,20 @@
   (let [options {:cursorLine line
                  :cursorX character
                  :prevCursorLine previous-line
-                 :prevCursorX previous-character}
+                 :prevCursorX previous-character
+                 :partialResult true}
         result (cljify (parinfer/indentMode text (jsify options)))]
     (jsify
      (if (:success result)
        {:success true
         :line (:cursorLine result)
         :character (:cursorX result)
-        :edits (editor/raplacement-edits-for-diffing-lines text (:text result))}
+        :new-text (:text result)
+        :edits (editor/replacement-edits-for-diffing-lines text (:text result))}
        {:success false
-        :error-msg (get-in result [:error :message])}))))
+        :error-msg (-> result :error :message)
+        :line (-> result :error :lineNo)
+        :character (-> result :error :x)}))))
 
 (defn infer-parens-bridge
   [^js m]
@@ -42,7 +46,8 @@
   "Calculate the edits needed for infering indents in `text`,
    and where the cursor should be placed to 'stay' in the right place."
   [{:keys [text line character previous-line previous-character changes]}]
-  (let [options {:cursorLine line :cursorX character
+  (let [options {:cursorLine line
+                 :cursorX character
                  :prevCursorLine previous-line
                  :prevCursorX previous-character
                  :changes (mapv (fn [change]
@@ -51,15 +56,60 @@
                                    :oldText (:old-text change)
                                    :newText (:new-text change)})
                                 changes)}
-        result (cljify (parinfer/smartMode text (jsify options)))]
+        result (cljify (parinfer/parenMode text (jsify options)))]
     (jsify
      (if (:success result)
        {:success true
         :line (:cursorLine result)
         :character (:cursorX result)
-        :edits (editor/raplacement-edits-for-diffing-lines text (:text result))}
+        :edits (editor/replacement-edits-for-diffing-lines text (:text result))}
        {:success false
-        :error-msg (get-in result [:error :message])}))))
+        :error-msg (-> result :error :message)
+        :line (-> result :error :lineNo)
+        :character (-> result :error :x)}))))
+
+(comment
+  (def broken-indentation "
+(ns foo
+  (:require [clojure.string :refer [blank?
+                                    lower-case
+                                    split
+                                  trim]])) ;; note: incorrect indentation
+
+(def foo)
+")
+
+  (infer-indents {:text broken-indentation
+                  :line 7
+                  :character 8})
+
+  (def good-indentation "
+(ns foo
+  (:require [clojure.string :refer [blank?
+                                    lower-case
+                                    split
+                                    trim]])) ;; note: correct indentation
+
+(def foo)
+")
+
+  (infer-indents {:text good-indentation
+                  :line 7
+                  :character 8})
+
+  (def broken-structure "
+(ns foo
+  (:require [clojure.string :refer [blank?
+                                    lower-case
+                                    split
+                                    trim])) ;; note: broken structure
+
+(def foo)
+")
+  (infer-indents {:text broken-structure
+                  :line 7
+                  :character 8}))
+
 
 (defn infer-indents-bridge
   [^js m]