+++ src/cursor-doc/model.ts
@@ -1,6 +1,7 @@
 import { Scanner, Token, ScannerState } from "./clojure-lexer";
 import { LispTokenCursor } from "./token-cursor";
-import { deepEqual as equal} from "../util/object"
+import { deepEqual as equal } from "../util/object";
+import * as parinfer from '../calva-fmt/src/infer';
 
 let scanner: Scanner;
 
@@ -84,6 +85,8 @@ export type ModelEditOptions = {
     undoStopBefore?: boolean,
     formatDepth?: number,
     skipFormat?: boolean,
+    parensInferred?: boolean,
+    rangeFormatted?: boolean,
     selection?: ModelEditSelection
 };
 
@@ -96,7 +99,8 @@ export interface EditableModel {
      * @param edits
      */
     edit: (edits: ModelEdit[], options: ModelEditOptions) => Thenable<boolean>;
-
+    parinferReadiness: parinfer.ParinferReadiness,
+    isWritable: boolean;
     getText: (start: number, end: number, mustBeWithin?: boolean) => string;
     getLineText: (line: number) => string;
     getOffsetForLine: (line: number) => number;
@@ -120,6 +124,14 @@ export interface EditableDocument {
 export class LineInputModel implements EditableModel {
     /** How many characters in the line endings of the text of this model? */
     constructor(readonly lineEndingLength: number = 1, private document?: EditableDocument) { }
+
+    isWritable = true;
+    
+    parinferReadiness: parinfer.ParinferReadiness = {
+        isIndentationHealthy: false,
+        isStructureHealthy: false
+    }
+    performInferParens = true;
 
     /** The input lines. */
     lines: TextLine[] = [new TextLine("", this.getStateForLine(0))];