+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 Changes to Calva.
 
 ## [Unreleased]
+- Re-enable, Experimental: [Add Parinfer Mode](https://github.com/BetterThanTomorrow/calva/issues/253)
+- Re-enable, Experimental: [Add option to keep text more fully formatted as you type](https://github.com/BetterThanTomorrow/calva/issues/1406)
+- Re-enable, [Remove `calva.fmt.formatAsYouType` option](https://github.com/BetterThanTomorrow/calva/issues/1407)
 
 ## [2.0.232] - 2021-12-31
 - [Prevent warning during deps.edn jack-in](https://github.com/BetterThanTomorrow/calva/issues/1355)